+++ backend/canisters/group_index/api/src/main.rs
@@ -15,7 +15,7 @@ fn main() {
     generate_candid_method!(group_index, unfreeze_community, update);
     generate_candid_method!(group_index, add_hot_group_exclusion, update);
     generate_candid_method!(group_index, remove_hot_group_exclusion, update);
-    generate_candid_method!(group_index, set_community_moderation_tags, update);
+    generate_candid_method!(group_index, set_community_moderation_flags, update);
     generate_candid_method!(group_index, set_community_upgrade_concurrency, update);
     generate_candid_method!(group_index, set_group_upgrade_concurrency, update);