+++ backend/canisters/group_index/impl/src/model/moderation_flags.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 bitflags! {
     #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, Serialize, Deserialize)]
-    pub struct ModerationTags: u32 {
+    pub struct ModerationFlags: u32 {
         const OFFENSIVE = 0b00000001;
         const ADULT     = 0b00000010;
     }