+++ backend/canisters/group_index/impl/src/model/public_communities.rs
@@ -1,4 +1,4 @@
-use crate::model::moderation_tags::ModerationTags;
+use crate::model::moderation_flags::ModerationFlags;
 use crate::model::private_communities::PrivateCommunityInfo;
 use crate::MARK_ACTIVE_DURATION;
 use search::{Document, Query};
@@ -44,7 +44,7 @@ impl PublicCommunities {
     pub fn search(
         &self,
         search_term: Option<String>,
-        exclude_moderation_tags: ModerationTags,
+        exclude_moderation_flags: ModerationFlags,
         page_index: u32,
         page_size: u8,
     ) -> Vec<CommunityMatch> {
@@ -53,7 +53,7 @@ impl PublicCommunities {
         let mut matches: Vec<_> = self
             .iter()
             .filter(|c| !c.is_frozen())
-            .filter(|c| !c.moderation_tags().intersects(exclude_moderation_tags))
+            .filter(|c| !c.moderation_flags().intersects(exclude_moderation_flags))
             .map(|c| {
                 let score = if let Some(query) = &query {
                     let document: Document = c.into();
@@ -131,7 +131,7 @@ pub struct PublicCommunityInfo {
     hotness_score: u32,
     gate: Option<AccessGate>,
     #[serde(default)]
-    moderation_tags: ModerationTags,
+    moderation_flags: ModerationFlags,
 }
 
 pub enum UpdateCommunityResult {
@@ -162,7 +162,7 @@ impl PublicCommunityInfo {
             activity: None,
             hotness_score: 0,
             frozen: None,
-            moderation_tags: ModerationTags::default(),
+            moderation_flags: ModerationFlags::default(),
         }
     }
 
@@ -211,12 +211,12 @@ impl PublicCommunityInfo {
         self.hotness_score = hotness_score;
     }
 
-    pub fn moderation_tags(&self) -> &ModerationTags {
-        &self.moderation_tags
+    pub fn moderation_flags(&self) -> &ModerationFlags {
+        &self.moderation_flags
     }
 
-    pub fn set_moderation_tags(&mut self, tags: ModerationTags) {
-        self.moderation_tags = tags;
+    pub fn set_moderation_flags(&mut self, flags: ModerationFlags) {
+        self.moderation_flags = flags;
     }
 }
 
@@ -231,7 +231,7 @@ impl From<&PublicCommunityInfo> for CommunityMatch {
             member_count: community.activity.as_ref().map_or(0, |a| a.member_count),
             channel_count: community.activity.as_ref().map_or(0, |a| a.channel_count),
             gate: community.gate.clone(),
-            moderation_tags: community.moderation_tags.bits(),
+            moderation_flags: community.moderation_flags.bits(),
         }
     }
 }