+++ backend/canisters/group_index/impl/src/queries/explore_communities.rs
@@ -1,4 +1,4 @@
-use crate::{model::moderation_tags::ModerationTags, read_state, RuntimeState};
+use crate::{model::moderation_flags::ModerationFlags, read_state, RuntimeState};
 use group_index_canister::explore_communities::{Response::*, *};
 use ic_cdk_macros::query;
 
@@ -21,16 +21,16 @@ fn explore_communities_impl(args: Args, state: &RuntimeState) -> Response {
         }
     }
 
-    let exclude_moderation_tags = match ModerationTags::from_bits(args.exclude_moderation_tags.unwrap_or_default()) {
+    let exclude_moderation_flags = match ModerationFlags::from_bits(args.exclude_moderation_flags.unwrap_or_default()) {
         Some(tags) => tags,
-        None => return InvalidTags,
+        None => return InvalidFlags,
     };
 
     let matches =
         state
             .data
             .public_communities
-            .search(args.search_term, exclude_moderation_tags, args.page_index, args.page_size);
+            .search(args.search_term, exclude_moderation_flags, args.page_index, args.page_size);
 
     Success(SuccessResult { matches })
 }