+++ backend/canisters/group_index/impl/src/updates/set_community_moderation_flags.rs
@@ -1,18 +1,17 @@
-use crate::{model::moderation_tags::ModerationTags, mutate_state, read_state, RuntimeState};
+use crate::{model::moderation_flags::ModerationFlags, mutate_state, read_state, RuntimeState};
 use candid::Principal;
 use canister_tracing_macros::trace;
-use group_index_canister::set_community_moderation_tags::{Response::*, *};
+use group_index_canister::set_community_moderation_flags::{Response::*, *};
 use ic_cdk_macros::update;
 use types::CanisterId;
 use user_index_canister_c2c_client::{lookup_user, LookupUserError};
 
 #[update]
 #[trace]
-async fn set_community_moderation_tags(args: Args) -> Response {
+async fn set_community_moderation_flags(args: Args) -> Response {
     let PrepareResult {
         caller,
         user_index_canister_id,
-        ..
     } = match read_state(|state| prepare(&args, state)) {
         Ok(result) => result,
         Err(response) => return response,
@@ -34,12 +33,12 @@ struct PrepareResult {
 
 fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareResult, Response> {
     if let Some(community) = state.data.public_communities.get(&args.community_id) {
-        if args.tags == community.moderation_tags().bits() {
+        if args.flags == community.moderation_flags().bits() {
             return Err(Unchanged);
         }
 
-        if ModerationTags::from_bits(args.tags).is_none() {
-            return Err(InvalidTags);
+        if ModerationFlags::from_bits(args.flags).is_none() {
+            return Err(InvalidFlags);
         }
 
         Ok(PrepareResult {
@@ -53,7 +52,7 @@ fn prepare(args: &Args, state: &RuntimeState) -> Result<PrepareResult, Response>
 
 fn commit(args: &Args, state: &mut RuntimeState) -> Response {
     if let Some(community) = state.data.public_communities.get_mut(&args.community_id) {
-        community.set_moderation_tags(ModerationTags::from_bits(args.tags).unwrap());
+        community.set_moderation_flags(ModerationFlags::from_bits(args.flags).unwrap());
         Success
     } else {
         CommunityNotFound