+++ backend/canisters/group_index/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Public community moderation tags ([#3911](https://github.com/open-chat-labs/open-chat/pull/3911))
+- Public community moderation flags ([#3911](https://github.com/open-chat-labs/open-chat/pull/3911))
 
 ### Changed